Extract minifed CSS and JS from the index.php file
The minifed files are commited to the repo because one of the
requirement is to have it working out of the box.

The developers need to:

* install yarn
* run: yarn install
* run: grunt

This needs to be done everytime the unminified version is changed.

Also remove files in .gitignore that should be commited.
And split the CSS for admin page in its own file.